+++ ui/mirage/config.js
@@ -14,6 +14,8 @@ export function findLeader(schema) {
 export default function() {
   this.timing = 0; // delay for each request, automatically set to 0 during testing
 
+  this.logging = window.location.search.includes('mirage-logging=true');
+
   this.namespace = 'v1';
   this.trackRequests = Ember.testing;